SafeSelfTriggered

Documentation for SafeSelfTriggered.

SafeSelfTriggered.ComplexZonotope

SafeSelfTriggered.OfflineComputationOptions

SafeSelfTriggered.OfflineComputationOptions

SafeSelfTriggered.OfflineQuantities

SafeSelfTriggered.OfflineQuantities

SafeSelfTriggered.SelfTriggeredLinearControl

SafeSelfTriggered.SelfTriggeredLinearImpulsive

SafeSelfTriggered.SolverOptions

Base.:*

Base.rand

SafeSelfTriggered.bloat_input

SafeSelfTriggered.compare_computation_times

SafeSelfTriggered.error_scale_sequence

SafeSelfTriggered.fast_inclusion_scale

SafeSelfTriggered.inclusion_scale

SafeSelfTriggered.large_invariant_zero_centered

SafeSelfTriggered.state_scale_sequence

SafeSelfTriggered.trigger_time_upper_bound

SafeSelfTriggered.verify

SafeSelfTriggered.ComplexZonotope — Type

Type: ComplexZonotope{N<:Real, M<:Number}(center::Vector{N}, generators::Matrix{M}, scale_vector::Vector{N})

Complex Zonotope is a set representation for some convex sets containing complex valued vectors.
It is an extension of the simple zonotope representation to complex numbers.
Its real projection can however be a non-polytopic set, unlike a simple zonotope which is polytopic.

Let c be a real vector, G be a complex matrix and s be a positive real vector. Then
ComplexZonotope(c,G,s) = Set{c + G\zeta: \zeta is complex vector, |\zeta| <= s}

Constructor: ComplexZonotope(center,generators,scale_vector)

SafeSelfTriggered.OfflineComputationOptions — Type

SafeSelfTriggered.OfflineComputationOptions — Method

SafeSelfTriggered.OfflineQuantities — Type

struct OfflineQuantities
    time_step ::Real
    invariant ::ComplexZonotope{<:Real, <:Number}
    scaling_matrix ::Matrix{<:Number}
    state_scales ::Vector{<:Real}
    error_scales ::Vector{<:Real}        
end

Quantities required to be computed offline.

fields

time_step : time step for computing successive reachable sets

invariant : invariant complex zonotope used as template for encoding reachable sets

scaling_matrix : matrix which maps a point to the maximum scale inside the complex zonotope

state_scales : sequence of reachable sets encoded by scalars, obtained by successive action of state action matrix

error_scales : sequence of disturbance reachable sets encoded by scalars

SafeSelfTriggered.OfflineQuantities — Method

OfflineQuantities(L::SelfTriggeredLinearControl{N}, Zinit::ComplexZonotope{N,M}, T::Matrix{N}, solver_options::SolverOptions=SolverOptions()) where {N<:Real, M<:Number} -> ::OfflineQuantitites

offline pre-computation of quantities required for computing latter the online triggering time upper bound.

SafeSelfTriggered.SelfTriggeredLinearControl — Type

struct SelfTriggeredLinearControl{N<:Real}
    A::Matrix{N} # state action matrix
    B::Matrix{N} # control action matrix 
    E::Matrix{N} # disturbance action matrix 
    K::Matrix{N} # feedback gain matrix 
    tmin::N # minimum time elapsed before trigger
end

Specification of Self Triggered Linear Control System

SafeSelfTriggered.SelfTriggeredLinearImpulsive — Type

Description of system:
(t_i)_{i=0}^{\infty}: Sequence of impulse times
t_{i+1}-t_{i} \in [tau{\min},tau{\max}]
x(t): state of system at time t
dx(t)/dt = Ax(t) + Ev(t) : \maxi |v(t)|i \leq 1, if t\in [t_i,t_{i+1})
x(t_{i}) = Rt_{i}^{-} if i>1

Code macros:
"A" -> A, "R"->R, "E" -> E, "\mu" -> mu, "tau{\max}" -> tmax, "tau{\min}" -> tmin

Type: SelftriggeredLinearImpusive{N<:Number}[A::Matrix{N}, R::Matrix{N}, E::Matrix{N}, mu:Vector{N}, tmax::N, tmin::N]

Constructor:
SelftriggeredLinearImpulsive(A, T, E, tmin)

SafeSelfTriggered.SolverOptions — Type

mutable struct SolverOptions
    sequence_length::Real
    order_invariant_zonotope::Integer
    order_error_zonotope::Integer
    optimizer::Function
    function SolverOptions(; sequence_length::Integer=100, order_invariant_zonotope::Integer=2, order_error_zonotope::Integer=10, optimizer::Function=Mosek.Optimizer)
        return new(sequence_length, order_invariant_zonotope, order_error_zonotope, optimizer)
    end
end

contains options for computing offline the scalarized reachability sequences

fields

sequence_length : length of reachability sequences, default = 100

order_invariant_zonotope : order of the invariant zonotope, default = 2

order_error_zonotope : order of error zonotope used during computation, default = 10

optimizer : Optimization solver used, default : Mosek.optimizer

constructor

SolverOptions(; sequence_length::Integer=100, order_invariant_zonotope::Integer=2, order_error_zonotope::Integer=10, optimizer::Function=Mosek.Optimizer)

Base.:* — Method

A*Z: where A::Matrix{L}, Z::ComplexZonotope{N,M}

computes the complex zonotope obtained by pre-multiplying the complex zonotope Z with A.

Base.rand — Method

rand(::ComplexZonotope, dim::Integer, numgen::Integer)

computes a random complex zonotope of dimension <dim> and number of generators <numgen>.

SafeSelfTriggered.bloat_input — Method

SafeSelfTriggered.compare_computation_times — Method

SafeSelfTriggered.error_scale_sequence — Method

error_scale_sequence(L::SelfTriggeredLinearImpulsive, Zinv::ComplexZonotope{N,M}, t::Real, l::Integer; order = 10) where {N<:Real, M<:Number}

computes sequence of contractions required for inclusion inside an invariant complex zonotope of a sequence of disturbance input sets involved in computing reachable sets in successive time intervals.

SafeSelfTriggered.fast_inclusion_scale — Method

fast_inclusion_scale(Z1::ComplexZonotope{N1, M1}, Z2::ComplexZonotope{N2,M2}) where {N1<:Real, N2<:Real, M1<:Number, M2<:Number}

function computes the minimum amount of expansion required for a complex zonotope to contain another complex zonotope.

Let Z1::ComplexZonotope(G1,c1,s1), Z2::ComplexZonotope(G2,c2,s2)
Then inclusion_scale(Z1,c1,s1) = min lambda \in \reals_{>=0} s.t.
Z2 \subseteq ComplexZonotope(G1,c1,lambda*s1)

SafeSelfTriggered.inclusion_scale — Method

inclusion_scale(Z1::ComplexZonotope{N1, M1}, Z2::ComplexZonotope{N2,M2}) where {N1<:Real, N2<:Real, M1<:Number, M2<:Number}

function computes the minimum amount of expansion required for a complex zonotope to contain another complex zonotope.

Let Z1::ComplexZonotope(G1,c1,s1), Z2::ComplexZonotope(G2,c2,s2)
Then inclusion_scale(Z1,c1,s1) = min lambda \in \reals_{>=0} s.t.
Z2 \subseteq ComplexZonotope(G1,c1,lambda*s1)

SafeSelfTriggered.large_invariant_zero_centered — Method

SafeSelfTriggered.state_scale_sequence — Method

state_scale_sequence(L::SelfTriggeredLinearImpulsive, Zinv::ComplexZonotope{N,M}, t::Real, l::Integer) where {N<:Real, M<:Number}

computes sequence of contractions required for inclusion inside an invariant complex zonotope after transformation by a sequence of sets of transformation operators as {exp(A tau) | (i-1)t <= tau <= itau}_{i=1}^{infinity}.

SafeSelfTriggered.trigger_time_upper_bound — Method

trigger_time_upper_bound(L::SelfTriggeredLinearImpulsive{<:Real}, x::Vector{<:Real}, inctesstateaction::Matrix{<:Number}, t::Real, statescales::Vector{<:Real}, errorscales::Vector{<:Real})

computes the upper bound on the time for next impulse

SafeSelfTriggered.verify — Method